api/requirements.txt
# Core
pytz==2018.7
Django==2.1.3
django-configurations==2.1
gunicorn==19.9.0
newrelic==4.6.0.106

# For the persistence stores
psycopg2-binary==2.7.6.1
dj-database-url==0.5.0

# Model Tools
django-model-utils==3.1.2
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.9.0
Markdown==3.0.1
django-filter==2.0.0

# Developer Tools
ipdb==0.11
ipython==7.1.1
mkdocs==0.17.5
flake8==3.6.0

# Testing
mock==2.0.0
factory-boy==2.11.1
django-nose==1.4.6
nose-progressive==1.5.1
coverage==4.5.2

# Static and Media Storage
django-storages==1.7.1


